requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.1.0
celery==5.3.4
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
contourpy==1.1.1
cycler==0.11.0
Django==4.2.5
djangorestframework==3.14.0
download==0.3.5
fonttools==4.42.1
idna==3.4
joblib==1.3.2
kiwisolver==1.4.5
kombu==5.3.2
libsvmdata==0.4.1
matplotlib==3.8.0
numpy==1.26.0
packaging==23.1
pandas==2.1.1
Pillow==10.0.1
prompt-toolkit==3.0.39
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
rabbitmq-server==0.0.1
redis==5.0.1
requests==2.31.0
scikit-learn==1.3.1
scipy==1.11.2
seaborn==0.12.2
six==1.16.0
sqlparse==0.4.4
threadpoolctl==3.2.0
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.5
vine==5.0.0
wcwidth==0.2.6
xarray==2023.9.0
